Assess pattern of missing data

Pass a data frame and the missing pattern of all columns in the data frame. The data frame is returned unmodified.

count_na(data)

Arguments

data
data frame

Examples

trial %>% count_na()

#> [1] "TRUE = 'Available', FALSE = 'Not Available'"
#> # A tibble: 4 x 9
#> # Groups:   trt, age, marker, stage, grade, response, death, ttdeath [4]
#>   trt   age   marker stage grade response death ttdeath     n
#>   <lgl> <lgl> <lgl>  <lgl> <lgl> <lgl>    <lgl> <lgl>   <int>
#> 1 TRUE  FALSE TRUE   TRUE  TRUE  TRUE     TRUE  TRUE        9
#> 2 TRUE  TRUE  FALSE  TRUE  TRUE  TRUE     TRUE  TRUE        9
#> 3 TRUE  TRUE  TRUE   TRUE  TRUE  FALSE    TRUE  TRUE        9
#> 4 TRUE  TRUE  TRUE   TRUE  TRUE  TRUE     TRUE  TRUE      173

#> # A tibble: 200 x 8
#>    trt       age marker stage grade response death ttdeath
#>    <chr>   <dbl>  <dbl> <fct> <fct>    <int> <int>   <dbl>
#>  1 Drug       23  0.16  T3    I            1     0   24   
#>  2 Drug        9  1.11  T4    III          1     1   14.9 
#>  3 Drug       31  0.277 T1    I            1     0   24   
#>  4 Placebo    46  2.07  T4    II           1     1   10.2 
#>  5 Drug       51  2.77  T2    II           0     1   20.0 
#>  6 Drug       39  0.613 T1    III          1     1   16.6 
#>  7 Drug       37  0.354 T4    II           1     1   12.2 
#>  8 Drug       32  1.74  T4    III          0     1    9.18
#>  9 Drug       31  0.144 T4    I            1     1   19.4 
#> 10 Placebo    34  0.205 T2    III         NA     1    8.65
#> # ... with 190 more rows