Table difference between two groups

Provided a stats::lm object, a table of adjusted differences are returned.

tbl_ancova(data, y, x, formula = "{y} ~ {x}", label = NULL,
  method.args = NULL, conf.level = 0.95, hide_n = FALSE,
  estimate_fun = style_sigfig, pvalue_fun = style_pvalue,
  method = stats::lm)

Arguments

data
Data frame to be used in univariate regression modeling. Data frame includes the outcome variable(s) and the independent variables.

y
vector of continuous outcome variables. One ANOVA/ANCOVA model will be computed for each outcome. All results returned in a single table.

x
string indicating the binary comparison variable

formula
String of the model formula. Uses glue::glue syntax. Default is "{y} ~ {x}", where {y} is the dependent variable, and {x} represents a single covariate. For a random intercept model, the formula may be formula = "{y} ~ {x} + (1 | gear)".

label
List of formulas specifying variables labels, e.g. list("age" ~ "Age, yrs", "ptstage" ~ "Path T Stage")

method.args
List of additional arguments passed on to the regression function defined by method.

conf.level
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

hide_n
Hide N column. Default is FALSE

estimate_fun
Function to round and format coefficient estimates. Default is style_sigfig when the coefficients are not transformed, and style_ratio when the coefficients have been exponentiated.

pvalue_fun
Function to round and format p-values. Default is style_pvalue. The function must have a numeric vector input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, purrr::partial(style_pvalue, digits = 2)).

method
Regression method (e.g. lm, glm, survival::coxph, and more).

Examples

trial %>%
  tbl_ancova(y = c("age", "marker"), x = "trt")

#> Error in tbl_summary_input_checks(data, by, label, type, value, statistic,     digits, missing, missing_text, group, sort): 'by' not a column in 'data'.