Table difference between two groups

Provided a stats::lm object, a table of adjusted differences are returned.

tbl_ancova(data, y, x, formula = "{y} ~ {x}", label = NULL,
  method.args = NULL, conf.level = 0.95, hide_n = FALSE,
  estimate_fun = style_sigfig, pvalue_fun = style_pvalue,
  method = lm)

Arguments

data
data frame to be used in univariate regression modeling. Data frame includes the outcome variable(s) and the independent variables.

y
model outcome as a string (e.g. y = recurrence or y = Surv(time, recur))

x
string indicating the binary comparison variable

formula
string that becomes the model formula. Uses glue syntax. Default is "{y} ~ {x}", where {y} is the dependent variable, and {x} represents a single covariate. For a random intercept, the formula may be formula = "{y} ~ {x} + (1 | gear)".

label
list of variable labels to override default labels in the table output, e.g. list(age60 = "Age > 60").

method.args
list of additional arguments passed on to the regression function defined by method.

conf.level
must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

hide_n
hide N column. Default is FALSE

estimate_fun
function to round and format beta coefficient estimates. Default is style_sigfig when the coefficients are not transformed, and style_ratio when the coefficients have been exponentiated.

pvalue_fun
function to round and format p-values. Default is style_pvalue. The function must have a numeric vector input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, purrr::partial(style_pvalue, digits = 2)).

method
regression method (e.g. lm, glm, coxph, and more).

Examples

trial %>%
  tbl_ancova(y = c("age", "marker"), x = "trt")

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #zbfsciexrq .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #zbfsciexrq .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #zbfsciexrq .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #zbfsciexrq .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #zbfsciexrq .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #zbfsciexrq .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #zbfsciexrq .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #zbfsciexrq .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #zbfsciexrq .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* row_group.background.color */
#>   font-size: 16px;
#>   /* row_group.font.size */
#>   font-weight: initial;
#>   /* row_group.font.weight */
#>   border-top-style: solid;
#>   /* row_group.border.top.style */
#>   border-top-width: 2px;
#>   /* row_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* row_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* row_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* row_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* row_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #zbfsciexrq .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* row_group.background.color */
#>   font-size: 16px;
#>   /* row_group.font.size */
#>   font-weight: initial;
#>   /* row_group.font.weight */
#>   border-top-style: solid;
#>   /* row_group.border.top.style */
#>   border-top-width: 2px;
#>   /* row_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* row_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* row_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* row_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* row_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #zbfsciexrq .gt_striped {
#>   background-color: #f2f2f2;
#> }
#> 
#> #zbfsciexrq .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #zbfsciexrq .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #zbfsciexrq .gt_row {
#>   padding: 8px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #zbfsciexrq .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #zbfsciexrq .gt_summary_row {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 8px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #zbfsciexrq .gt_grand_summary_row {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* grand_summary_row.background.color */
#>   padding: 8px;
#>   /* grand_summary_row.padding */
#>   text-transform: inherit;
#>   /* grand_summary_row.text_transform */
#> }
#> 
#> #zbfsciexrq .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #zbfsciexrq .gt_first_grand_summary_row {
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #zbfsciexrq .gt_table_body {
#>   border-top-style: solid;
#>   /* table_body.border.top.style */
#>   border-top-width: 2px;
#>   /* table_body.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table_body.border.top.color */
#>   border-bottom-style: solid;
#>   /* table_body.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* table_body.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* table_body.border.bottom.color */
#> }
#> 
#> #zbfsciexrq .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #zbfsciexrq .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #zbfsciexrq .gt_center {
#>   text-align: center;
#> }
#> 
#> #zbfsciexrq .gt_left {
#>   text-align: left;
#> }
#> 
#> #zbfsciexrq .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #zbfsciexrq .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #zbfsciexrq .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #zbfsciexrq .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #zbfsciexrq .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #zbfsciexrq .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="zbfsciexrq" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>N</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Placebo</strong>, N = 93<sup class='gt_footnote_glyph'>1</sup></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Drug</strong>, N = 107<sup class='gt_footnote_glyph'>1</sup></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Difference</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>95% CI</strong><sup class='gt_footnote_glyph'>2</sup></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>p-value</strong></th>
#> </tr>
#> <tbody class='gt_table_body'>
#> <tr>
#> <td class='gt_row gt_left'>Age, yrs</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'>46 (13)</td>
#> <td class='gt_row gt_center'>48 (15)</td>
#> <td class='gt_row gt_center'>-1.9</td>
#> <td class='gt_row gt_center'>-6.0, 2.2</td>
#> <td class='gt_row gt_center'>0.4</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center gt_striped'>191</td>
#> <td class='gt_row gt_center gt_striped'>0.96 (0.83)</td>
#> <td class='gt_row gt_center gt_striped'>0.90 (0.89)</td>
#> <td class='gt_row gt_center gt_striped'>0.06</td>
#> <td class='gt_row gt_center gt_striped'>-0.18, 0.31</td>
#> <td class='gt_row gt_center gt_striped'>0.6</td>
#> </tr>
#> </tbody>
#> <tfoot>
#> <tr>
#> <td colspan='7' class='gt_footnote'><sup class='gt_footnote_glyph'><em>1</em></sup> Mean (Standard Deviation)<br /><sup class='gt_footnote_glyph'><em>2</em></sup> CI = Confidence Interval</td>
#> </tr>
#> </tfoot></table>
#> <!--gt table end-->
#> </div>